refactored

+++ app/src/main/java/com/kay/prog/easygift/data/models/Wish.kt
@@ -5,5 +5,5 @@ data class Wish (
     var nickname: String,
     var description: String,
     var webpage: String?,
-    var price: Double?
+    var price: Int?
 )
